PostNewTriggerTestsBase is added.
